// components/RecruitmentsAndOnboarding/JobRequisitions/types.ts
export interface JobRequisitions  {
    title: string;
    post_type: string;
    requisition_type: string;
    justification: string;
    number_of_positions: number;
    description: string;
    requirements: string;
    closing_date: string;
    proposed_salary_range:number;
    date_requested: string;
    status: string;
    date_created: string;
    date_updated: string;

}